requirements.txt
chromadb==1.1.1
google-genai==1.41.0
groq==0.32.0
keybert==0.9.0
langchain==0.3.27
langchain_community==0.3.31
langchain_core==0.3.78
langchain_google_genai==2.1.12
langdetect==1.0.9
langgraph==0.6.9
langsmith==0.4.33
playwright==1.55.0
protobuf==6.32.1
pydantic==2.12.0
python-dotenv==1.1.1
reportlab==4.4.4
sentence_transformers==5.1.1
streamlit==1.50.0
transformers==4.57.0
torch==2.8.0
sentencepiece==0.2.1
typing_extensions==4.15.0
